Changes to the fundamental base arby is main file gui controls GUI binanceapi controls the Binance API
