+++ gui.py
@@ -1,0 +1,6 @@
+from tkinter import *
+
+def start():
+    root = Tk()
+
+    root.mainloop()
